requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.0
email-validator==2.1.0
qrcode[pil]==7.4.2
gspread==5.12.0
google-auth==2.26.1